refactor(speakers): decompose SpeakerPanel into focused subcomponents and hooks (FE-012)
